Production-Editor/Supra-Adjuster/src/requirements.txt
tk==0.1.0
docx2python==2.0.4
Pillow==10.0.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
